Extract macos titlebar into its own file
